Agents/requirements.txt
python-dotenv
requests
openai 
langchain
pydantic
fuzzywuzzy